pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gander-llm-cli"
version = "0.1.0"
description = "CLI that returns a human-facing LLM answer plus structured JSON fields."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer==0.12.5",
    "pydantic==2.8.2",
    "httpx==0.27.2",
    "beautifulsoup4==4.12.3",
    "tiktoken==0.7.0",
    "python-dotenv==1.0.1",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
gander-llm-cli = "gander_llm_cli.cli:main"
